Move tabs to util directory
