+++ src/sidebar/services/root-thread.js
@@ -4,7 +4,7 @@ const buildThread = require('../build-thread');
 const events = require('../events');
 const memoize = require('../util/memoize');
 const metadata = require('../util/annotation-metadata');
-const tabs = require('../tabs');
+const tabs = require('../util/tabs');
 const uiConstants = require('../ui-constants');
 
 function truthyKeys(map) {
